Fix commands & value objects
