--- Domain/ValueObjects/MakeTimeDeposit.php
@@ -1,7 +1,0 @@
-<?php
-
-
-class MakeTimeDeposit
-{
-
-}
