+++ app/Application/Commands/TimeDeposit/ReinvestTimeDepositCommand.php
@@ -20,57 +20,4 @@ class ReinvestTimeDepositCommand
         $this->approve = $approve;
     }
 
-    public function getFullName(): string
-    {
-        return $this->fullName;
-    }
-
-    public function getInitAmount(): float
-    {
-        return $this->amount;
-    }
-
-    /**
-     * @return bool
-     */
-    public function isApprove(): bool
-    {
-        return $this->approve;
-    }
-
-    /**
-     * @return float
-     */
-    public function getAmount(): float
-    {
-        return $this->amount;
-    }
-
-    public function getFinalBalance(): float
-    {
-        return $this->finalBalance;
-    }
-
-    public function getDays(): int
-    {
-        return $this->days;
-    }
-
-    public function __serialize(): array
-    {
-        return [
-            'amount' => $this->amount,
-            'finalBalance' => $this->finalBalance,
-            'fullName' => $this->fullName,
-            'days' => $this->days
-        ];
-    }
-
-    public function __toString()
-    {
-        return 'Nombre y apellido: ' . $this->fullName .
-            ', Monto inicial:' . $this->amount .
-            ', Balance Final: ' . $this->finalBalance .
-            ' , Duración: ' . $this->days;
-    }
 }
